fix: improve a base contract

+++ contracts/base/AccessControlExtUpgradeable.sol
@@ -40,7 +40,7 @@ abstract contract AccessControlExtUpgradeable is AccessControlUpgradeable {
      * @param accounts The accounts to grant the role to.
      */
     function grantRoleBatch(bytes32 role, address[] memory accounts) public virtual onlyRole(getRoleAdmin(role)) {
-        for (uint i = 0; i < accounts.length; i++) {
+        for (uint256 i = 0; i < accounts.length; i++) {
             _grantRole(role, accounts[i]);
         }
     }
@@ -56,7 +56,7 @@ abstract contract AccessControlExtUpgradeable is AccessControlUpgradeable {
      * @param accounts The accounts to revoke the role from.
      */
     function revokeRoleBatch(bytes32 role, address[] memory accounts) public virtual onlyRole(getRoleAdmin(role)) {
-        for (uint i = 0; i < accounts.length; i++) {
+        for (uint256 i = 0; i < accounts.length; i++) {
             _revokeRole(role, accounts[i]);
         }
     }
